requirements.txt
# This file is autogenerated by Prequ.  To update, run:
#
#   prequ update
#
-e .
numpy==1.14.5             # via opencv-python
opencv-python==3.4.2.17   # via meterelf
pyyaml==3.13              # via meterelf
